requirements.txt
rdap
mysql-connector-python
pandas
SQLAlchemy
streamlit